# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Assistant Messenger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Assistant {
  id: number;
  name: string;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  avatarEmoji?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isPinned: boolean;
  isArchived: boolean;
  /** @nullable */
  lastMessageAt?: string | null;
  /** @nullable */
  lastMessagePreview?: string | null;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface AssistantInput {
  /** @minLength 1 */
  name: string;
  instructions?: string;
  avatarEmoji?: string;
  avatarUrl?: string;
}

export interface AssistantUpdate {
  /** @minLength 1 */
  name?: string;
  instructions?: string;
  avatarEmoji?: string;
  avatarUrl?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  assistantId: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface AvatarGenerateInput {
  prompt: string;
}

export interface AvatarOutput {
  b64_json: string;
}

export interface SearchMessage {
  id: number;
  assistantId: number;
  assistantName: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface SearchResults {
  assistants: Assistant[];
  messages: SearchMessage[];
}

export interface ApiError {
  error: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiVoiceMessageInput {
  /** Base64-encoded audio data */
  audio: string;
}

export type SearchAssistantsParams = {
q: string;
};