# lib/api-zod/src/generated/types/assistant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Assistant Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface Assistant {
  id: number;
  name: string;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  avatarEmoji?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isPinned: boolean;
  isArchived: boolean;
  /** @nullable */
  lastMessageAt?: Date | null;
  /** @nullable */
  lastMessagePreview?: string | null;
  messageCount: number;
  createdAt: Date;
  updatedAt: Date;
}