# lib/api-zod/src/generated/types/assistantInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Assistant Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantInput {
  /** @minLength 1 */
  name: string;
  instructions?: string;
  avatarEmoji?: string;
  avatarUrl?: string;
}