# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Assistant Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageRole } from './messageRole';

export interface Message {
  id: number;
  assistantId: number;
  role: MessageRole;
  content: string;
  createdAt: Date;
}