# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personal Assistant Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { Assistant } from './assistant';
import type { SearchMessage } from './searchMessage';

export interface SearchResults {
  assistants: Assistant[];
  messages: SearchMessage[];
}